+++ packages/design-system/src/icons/fill/components/index.tsx
@@ -4,5 +4,5 @@ export { SvgCheck } from './Check';
 export { SvgErrorFill } from './ErrorFill';
 export { SvgLikeFill } from './LikeFill';
 export { SvgInstagramFill } from './InstagramFill';
-export { SvgTwitterFill } from './Twitter_fill';
+export { SvgTwitterFill } from './TwitterFill';
 export { SvgArrowRight } from './ArrowRight';
